removed auth, added bryce mention interaction
